+++ scripts/orchestrator/README.md
@@ -101,10 +101,11 @@ Implementation Snippet (deterministic_loop.py)
 import subprocess
 import json
 import time
+import os
+import sys
 
-def run_cli(args):
-    result = subprocess.run(["python3", "boomtick-pkg/cli/dev_tools/td_cli.py"] + args, capture_output=True, text=True)
-    return result.stdout.strip()
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from utils import run_cli
 
 def deterministic_routing_loop(poll_interval=3600):
     while True:
@@ -163,18 +164,11 @@ Implementation Snippet (continuous_dev_loop.py)
 import subprocess
 import json
 import time
+import os
+import sys
 
-def run_cli(args):
-    result = subprocess.run(["python3", "boomtick-pkg/cli/dev_tools/td_cli.py"] + args, capture_output=True, text=True)
-    return result.stdout.strip()
-
-def get_session_id():
-    # Helper to parse active session ID
-    pass
-
-def wait_for_agent(session_id):
-    # Polling logic from the main architecture
-    pass
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from utils import run_cli, get_session_id, wait_for_agent
 
 def execute_continuous_dev_loop(issue_queue):
     """
@@ -264,13 +258,13 @@ Dispatch Intent: Sends high-level JSON payloads, allowing Jules to autonomously 
 
 Orchestrator Script (agent_2_orchestrator.py)
 
-import subprocess
 import json
-import re
-import time
 import argparse
+import sys
+import os
 
-CLI_BASE = ["python3", "boomtick-pkg/cli/dev_tools/td_cli.py"]
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from utils import run_cli, get_session_id, wait_for_agent
 
 JULES_SYSTEM_PROMPT = """You are Jules, an autonomous software engineering agent. You execute actions on demand when given a task payload. You do not manage loops, and you do not schedule your own future executions. However, you are strictly responsible for verifying that you do not perform actions on the same asset too frequently.
 
@@ -286,37 +280,6 @@ Outcome B (Proceed): { "status": "SUCCESS", "timestamp": "...", "target_id": "..
 
 Initial Task Payload:
 """
-
-def run_cli(args):
-    """Executes a BoomTick CLI command and returns the standard output."""
-    cmd = CLI_BASE + args
-    try:
-        result = subprocess.run(cmd, capture_output=True, text=True, check=True)
-        return result.stdout.strip()
-    except subprocess.CalledProcessError as e:
-        print(f"CLI Error: {e.stderr}")
-        return None
-
-def get_session_id():
-    """Runs 'agent sync' and parses the output for the active session ID."""
-    stdout = run_cli(["agent", "sync"])
-    if not stdout:
-        return None
-    match = re.search(r"(?:Session ID|id):\s*([a-zA-Z0-9_-]+)", stdout, re.IGNORECASE)
-    return match.group(1) if match else None
-
-def wait_for_agent(session_id, poll_interval=10):
-    """Blocks execution until the agent reaches a terminal state or requests input."""
-    print(f"Polling state for session {session_id}...")
-    while True:
-        messages = run_cli(["agent", "messages", session_id])
-        if messages:
-            is_completed = "SUCCESS" in messages or "ABORTED_THROTTLED" in messages
-            is_waiting = "waiting for input" in messages.lower() or "failed" in messages.lower()
-            if is_completed or is_waiting:
-                print("Agent 1 is ready.")
-                break
-        time.sleep(poll_interval)
 
 def execute_orchestrator_loop(existing_session_id=None):
     """Main execution flow for Agent 2 orchestrating Agent 1 (Jules)."""
