+++ scripts/orchestrator/agent_2_orchestrator.py
@@ -1,11 +1,10 @@
-import subprocess
 import json
-import re
-import time
 import argparse
+import sys
 import os
 
-CLI_BASE = ["python3", "boomtick-pkg/cli/dev_tools/td_cli.py"]
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from utils import run_cli, get_session_id, wait_for_agent
 
 JULES_SYSTEM_PROMPT = """You are Jules, an autonomous software engineering agent. You execute actions on demand when given a task payload. You do not manage loops, and you do not schedule your own future executions. However, you are strictly responsible for verifying that you do not perform actions on the same asset too frequently.
