+++ scripts/orchestrator/experiments/genai_orchestrator.py
@@ -1,24 +1,11 @@
-import subprocess
 import json
 import os
+import sys
 from datetime import datetime
 
-CLI_BASE = ["python3", "boomtick-pkg/cli/dev_tools/td_cli.py"]
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+from utils import run_cli
 
-def run_cli(args):
-    """Executes a BoomTick CLI command and returns the standard output."""
-    cmd = CLI_BASE + args
-    env = os.environ.copy()
-    existing_path = env.get("PYTHONPATH", "")
-    local_paths = "boomtick-pkg/cli:boomtick-pkg/cli/dev_tools"
-    env["PYTHONPATH"] = f"{local_paths}:{existing_path}" if existing_path else local_paths
-    try:
-        result = subprocess.run(cmd, capture_output=True, text=True, check=True, env=env)
-        return result.stdout.strip()
-    except subprocess.CalledProcessError as e:
-        print(f"CLI Error: {e.stderr}")
-        return None
-import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "boomtick-pkg", "cli", "dev_tools")))
 try:
     from utils import call_ai, get_ai_synthesis_model, clean_llm_output
